refactor(tonk-schema): decouple operation types from interpretation
The operation data types — Application, Statement, Effect,
Conclusion, the mutation types — no longer reach into the schema's
query and resolution machinery. The Application -> QueryPlan lowering
is a free function application_to_plan in concept.rs (Application no
longer impls dialog_query::Application); the effect query machinery
(AnonymousRuleQuery, EffectByEntity, EffectsByOn, the Statement impl)
moves to a new effect_query module. effect.rs is left as the Effect
data type alone.

transact, effect, conclusion, mutation now import only each other
and the leaf meta module — the operation cluster is a clean leaf,
ready to extract into tonk-core.
